Sample project should now appear too
This sample project should be in the same folder as the application.
For Windows users that's the same folder as the ScenLang.exe file
And for Mac users that's the same folder where the ScenLang.app folder is put it. So don't put it in the app bundle folder itself.
And yeah, this should work in Linux too, and then it's the same folder where ScenLang binary file has been located.